all: make put.io plugin also a service
Kodi add-ons are a single shot python scripts, which means that Kodi
executes the script on 'enter a directory' or a 'play this item'. We
need to set last media position, which requires an event-loop mechanism
to observe Kodi player events.

Kodi 'services' solves this problem. It provides a Monitor mechanism to
poll the player status. Primitive, but still...

We are polling player status for every 10 seconds, and if certain
conditions are met, send the last video position Put.io API.

Fixes #6