main: use setResolvedUrl instead of xbmc.Player().play()
setResolvedUrl has all the magic ingredients when it comes to playing
video. xbmc.Player() class is a raw class and there are no context
available if the video is played before.

setResolvedUrl solves most of the player problems, such as gathering
mime-type on play time, resuming the video based on the user input etc.